修复歌单接口各种问题，获取用户喜欢音乐接口需要喜欢歌单放第一条

+++ api/src/main/java/org/api/neteasecloudmusic/model/vo/playlist/PlaylistItem.java
@@ -12,7 +12,7 @@ public class PlaylistItem {
 	private String description;
 	private Long trackNumberUpdateTime;
 	private boolean subscribed;
-	private Integer trackCount;
+	private Long trackCount;
 	private Integer adType;
 	private String coverImgIdStr;
 	private Integer specialType;
